Don't try to build the OpenGL viewer for MinGW.
Commit 018253cfc8a34805c4fc1e0f955e3696db60886a upstream, which added
building the OpenGL viewer for the MacOS target, also made it so that
'make' will try to build the viewer for all targets that the Makefile
is used for, including for MinGW. As there isn't support yet for
building the OpenGL viewer for MinGW, and default is to build it for
X11, 'make' fails and bails out at that point.

To work around this, exclude building mupdf-gl for the MinGW target.

At some later date support will be added for building the viewer for
MinGW.