moz: Create the 'mozilla' sub-directory by default.
The existing setup was causing compilation to fail, when built with
MinGW.org's native system (i.e. under MSYS on Windows), and having the
'mozilla' sub-directory created early on with the other platform
sub-directories (along with referring to it's path more explicitly in
the compilation rules) seems to work.

Also, while we're at it, add the NPAPI CFLAGS to the general CFLAGS
list.